fix: 修复 Windows 类型转换警告
- platform_channel.cpp: 修复 size_t 到 DWORD 的类型转换
- platform_channel.cpp: 修复 int64_t 到 EncodableValue 的类型转换
- mihomo_core.cpp: 修复 int 到 DWORD 的有符号/无符号类型转换
- mihomo_core.cpp: 修复 size_t 到 DWORD 的类型转换

+++ windows/runner/mihomo_core.cpp
@@ -372,7 +372,7 @@ std::string MihomoCore::HttpGet(const std::string& path) {
         if (!controllerSecret_.empty()) {
             std::wstring authHeader = L"Authorization: Bearer " +
                 std::wstring(controllerSecret_.begin(), controllerSecret_.end());
-            WinHttpAddRequestHeaders(hRequest, authHeader.c_str(), -1,
+            WinHttpAddRequestHeaders(hRequest, authHeader.c_str(), static_cast<DWORD>(-1),
                 WINHTTP_ADDREQ_FLAG_ADD);
         }
 
@@ -437,10 +437,10 @@ std::string MihomoCore::HttpPut(const std::string& path, const std::string& body
             headers += L"Authorization: Bearer " +
                 std::wstring(controllerSecret_.begin(), controllerSecret_.end()) + L"\r\n";
         }
-        WinHttpAddRequestHeaders(hRequest, headers.c_str(), -1, WINHTTP_ADDREQ_FLAG_ADD);
+        WinHttpAddRequestHeaders(hRequest, headers.c_str(), static_cast<DWORD>(-1), WINHTTP_ADDREQ_FLAG_ADD);
 
         if (!WinHttpSendRequest(hRequest, WINHTTP_NO_ADDITIONAL_HEADERS, 0,
-            (LPVOID)body.c_str(), body.length(), body.length(), 0)) {
+            (LPVOID)body.c_str(), static_cast<DWORD>(body.length()), static_cast<DWORD>(body.length()), 0)) {
             throw std::runtime_error("WinHttpSendRequest failed");
         }
